payment-done

+++ farmer-micro/src/main/java/farmer/service/FarmerService.java
@@ -9,10 +9,12 @@ import java.util.Map;
 import org.springframework.amqp.core.AmqpTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
 
+import farmer.com.util.JwtUtil;
 import farmer.dto.CropDto;
 import farmer.dto.EditDto;
 import farmer.dto.FarmerDto;
@@ -28,9 +30,13 @@ import farmer.repopsitory.BankAccountRepository;
 import farmer.repopsitory.CropRespository;
 import farmer.repopsitory.FarmEmailRepository;
 import farmer.repopsitory.FarmerRepository;
-
+import farmer.serviceInterface.FarmerServiceInterface;
+import lombok.extern.slf4j.Slf4j;
+@Slf4j
 @Service
-public class FarmerService {
+public class FarmerService implements FarmerServiceInterface {
+	@Autowired
+	private JwtUtil jwtUtil;
 	@Autowired
 	private AmqpTemplate rabbitTemplate;
 
@@ -53,6 +59,7 @@ public class FarmerService {
 	@Autowired
 	private FarmEmailRepository emailRepo;
 
+	@Override
 	public boolean register(FarmerDto farmerDto) {
 
 		Address address = new Address(farmerDto.getHouseNo(), farmerDto.getLocality(), farmerDto.getTown(),
@@ -68,8 +75,8 @@ public class FarmerService {
 		cropRepo.save(crops);
 
 		FarmerModel farmerModel = new FarmerModel(farmerDto.getFirstName(), farmerDto.getLastName(),
-				farmerDto.getEmail(), farmerDto.getContact(),"Active", farmerDto.getUserName(), farmerDto.getPassword(), address,
-				new ArrayList<>(), bankAccountDeatil, farmerDto.getRating());
+				farmerDto.getEmail(), farmerDto.getContact(), "Active", farmerDto.getUserName(),
+				farmerDto.getPassword(), address, new ArrayList<>(), bankAccountDeatil, farmerDto.getRating());
 
 		farmerModel.getCrops().add(crops);
 		crops.setFarmer(farmerModel);
@@ -80,60 +87,75 @@ public class FarmerService {
 
 		UserLogin login = new UserLogin(farmerDto.getUserName(), farmerDto.getPassword(), "farmer");
 		rabbitTemplate.convertAndSend(exchange, routingkey, login);
-		System.out.println("Send msg = " + login);
-
+		log.info("Send msg = " + login);
+		
 		SimpleMailMessage msg = new SimpleMailMessage();
 		List<FarmEmails> emails = emailRepo.findAll();
-		for (FarmEmails i : emails) {
+		for (FarmEmails i : emails) 
+		{
+			msg.setSubject("Farmer-Dealer-Web-Service");
+			msg.setText("Checkout new crops freshly from the farms of farmer registered on our site.");
 			msg.setTo(i.getEmails());
+			javaMailSender.send(msg);
+				log.info("mail sent");
 		}
-
-		msg.setSubject("Farmer-Dealer-Web-Service");
-		msg.setText("Checkout new crops freshly from the farms of farmer registered on our site.");
-
-		javaMailSender.send(msg);
 
 		return true;
 
 	}
 
-	public boolean addCrop(CropDto crop) {
-		FarmerModel farmer = repo.getById(crop.getUserName());
-		Crop crops = new Crop(crop.getId(), crop.getCropName(), crop.getCropType(), crop.getCropQuantity(),
-				crop.getPrice(), farmer);
-		cropRepo.save(crops);
-		farmer.getCrops().add(crops);
-		cropRepo.save(crops);
-
-		SimpleMailMessage msg = new SimpleMailMessage();
-		List<FarmEmails> emails = emailRepo.findAll();
-		for (FarmEmails i : emails) {
-			msg.setTo(i.getEmails());
-		}
-
-		msg.setSubject("Farmer-Dealer-Web-Service");
-		msg.setText("Checkout new crops freshly from the farms of farmer registered on our site.");
-
-		javaMailSender.send(msg);
-		return true;
-	}
-
-	public boolean removeCrop(String userName,Integer id) {
-
-		FarmerModel farmerModel = repo.findById(userName).get();
-		for (Crop crop : farmerModel.getCrops()) {
-			if (crop.getId() == id) {
-
-				farmerModel.getCrops().remove(crop);
-				cropRepo.delete(crop);
-				repo.save(farmerModel);
-
+	@Override
+	public boolean addCrop(CropDto crop, ServerHttpRequest request) {
+		if (validateToken(request)) {
+			
+			SimpleMailMessage msg = new SimpleMailMessage();
+			List<FarmEmails> emails = emailRepo.findAll();
+			for (FarmEmails i : emails) 
+			{
+				msg.setSubject("Farmer-Dealer-Web-Service");
+				msg.setText("Checkout new crops freshly from the farms of farmer registered on our site.");
+				msg.setTo(i.getEmails());
+				javaMailSender.send(msg);
+				log.info("mail sent");
 			}
-		}
+				
+				FarmerModel farmer = repo.getById(crop.getUserName());
+				Crop crops = new Crop(crop.getId(), crop.getCropName(), crop.getCropType(), crop.getCropQuantity(),
+						crop.getPrice(), farmer);
+				cropRepo.save(crops);
+				farmer.getCrops().add(crops);
+				cropRepo.save(crops);
 
-		return true;
+			
+
+			return true;
+		} else
+			log.info("invalid token provided");
+			throw new TokenValidationError("Token is not valid", org.springframework.http.HttpStatus.FORBIDDEN);
 	}
 
+	@Override
+	public boolean removeCrop(String userName, Integer id, ServerHttpRequest request) {
+
+		if (validateToken(request)) {
+			FarmerModel farmerModel = repo.findById(userName).get();
+			for (Crop crop : farmerModel.getCrops()) {
+				if (crop.getId() == id) {
+
+					farmerModel.getCrops().remove(crop);
+					cropRepo.delete(crop);
+					repo.save(farmerModel);
+
+				}
+			}
+
+			return true;
+		} else
+			log.info("invalid token provided");
+			throw new TokenValidationError("Token is not valid", org.springframework.http.HttpStatus.FORBIDDEN);
+	}
+
+	@Override
 	public List<Crop> getFarmerCrops() {
 		List<Crop> finalCrop = new ArrayList<>();
 		List<Crop> crop = cropRepo.findAll();
@@ -146,63 +168,73 @@ public class FarmerService {
 
 	}
 
+	@Override
 	public void removeFamer(String userName) {
 
-		FarmerModel farmerModel=repo.findById(userName).get();
+		FarmerModel farmerModel = repo.findById(userName).get();
 		farmerModel.setStatus("Inactive");
 		repo.save(farmerModel);
-		
 
 	}
 
+	@Override
 	public boolean rateFarmer(RatingDto dto) {
 
 		FarmerModel farmer;
-		farmer = repo.getById(dto.getUserId());
+		Crop crop = cropRepo.findById(dto.getCropId()).get();
+		farmer = repo.findById(crop.getFarmer().getUserName()).get();
 		farmer.setRating(dto.getRating());
 		repo.save(farmer);
 		return true;
 
 	}
 
-	public boolean editProfile(EditDto dto) {
-		FarmerModel farmer = new FarmerModel();
-		farmer = repo.findById(dto.getUserName()).get();
-		farmer.setFirstName(dto.getFirstName());
-		farmer.setLastName(dto.getLastName());
-		farmer.setEmail(dto.getEmail());
-		farmer.setContact(dto.getContact());
-		farmer.setUserName(dto.getUserName());
+	@Override
+	public boolean editProfile(EditDto dto, ServerHttpRequest request) {
+		if (validateToken(request)) {
+			FarmerModel farmer = new FarmerModel();
+			farmer = repo.findById(dto.getUserName()).get();
+			farmer.setFirstName(dto.getFirstName());
+			farmer.setLastName(dto.getLastName());
+			farmer.setEmail(dto.getEmail());
+			farmer.setContact(dto.getContact());
+			farmer.setUserName(dto.getUserName());
 
-		Address address = new Address(dto.getHouseNo(), dto.getLocality(), dto.getTown(), dto.getDistrict(),
-				dto.getState(), dto.getPostalCode(), null);
-		farmer.setAddress(address);
+			Address address = new Address(dto.getHouseNo(), dto.getLocality(), dto.getTown(), dto.getDistrict(),
+					dto.getState(), dto.getPostalCode(), null);
+			farmer.setAddress(address);
 
-		BankAccountDeatil bankAccountDeatil = new BankAccountDeatil(dto.getBankAccountHolderName(),
-				dto.getBankAccountNo(), dto.getIfscCode());
+			BankAccountDeatil bankAccountDeatil = new BankAccountDeatil(dto.getBankAccountHolderName(),
+					dto.getBankAccountNo(), dto.getIfscCode());
 
-		bankRepo.save(bankAccountDeatil);
-		farmer.setBankAccountDeatil(bankAccountDeatil);
-		repo.save(farmer);
-		address.setFarmer(farmer);
-		addressRepo.save(address);
+			bankRepo.save(bankAccountDeatil);
+			farmer.setBankAccountDeatil(bankAccountDeatil);
+			repo.save(farmer);
+			address.setFarmer(farmer);
+			addressRepo.save(address);
 
-		return true;
+			return true;
+		} else
+			log.info("invalid token provided");
+			throw new TokenValidationError("Token is not valid", org.springframework.http.HttpStatus.FORBIDDEN);
 	}
 
-	public Address getAddress(int  id) {
+	@Override
+	public Address getAddress(int id) {
 
-		Crop crop =cropRepo.findById(id).get();
+		Crop crop = cropRepo.findById(id).get();
 		return crop.getFarmer().getAddress();
 
 	}
 
+	@Override
 	public BankAccountDeatil getbankDetails(int id) {
 
-		Crop crop =cropRepo.findById(id).get();
+		Crop crop = cropRepo.findById(id).get();
 		return crop.getFarmer().getBankAccountDeatil();
 	}
 
+	@Override
 	public Boolean quantityManagement(HashMap<Integer, Integer> CropIds) {
 		Iterator<Map.Entry<Integer, Integer>> itr = CropIds.entrySet().iterator();
 		while (itr.hasNext()) {
@@ -211,43 +243,84 @@ public class FarmerService {
 			Crop crop = cropRepo.findById(entry.getKey()).get();
 			crop.setCropQuantity(crop.getCropQuantity() - entry.getValue());
 			cropRepo.save(crop);
-			
+
 		}
 		return null;
 	}
 
-	public Boolean saveFarmEmail(ArrayList<String> Emails) {
+	@Override
+	public Boolean saveFarmEmail(String Emails) {
 
-		for (String i : Emails) {
-			FarmEmails farmEmails = new FarmEmails(i);
-			emailRepo.save(farmEmails);
-
-		}
+		FarmEmails farmEmails = new FarmEmails(Emails);
+		emailRepo.save(farmEmails);
 
 		return true;
 	}
 
-	public FarmerDto getFarmerDetails(String userName) {
-		FarmerModel farmerModel = repo.findById(userName).get();
-		FarmerDto farmerDto = new FarmerDto(farmerModel.getFirstName(),farmerModel.getLastName(),farmerModel.getEmail(),
-				farmerModel.getContact(),farmerModel.getUserName(),farmerModel.getPassword(),farmerModel.getRating(),
-				farmerModel.getAddress().getHouseNo(),farmerModel.getAddress().getLocality(),farmerModel.getAddress().getTown(),farmerModel.getAddress().getDistrict(),
-				farmerModel.getAddress().getState(),farmerModel.getAddress().getPostalCode(),0,null,null,0,0,farmerModel.getBankAccountDeatil().getBankAccountNo(),
-				farmerModel.getBankAccountDeatil().getIfscCode(),farmerModel.getBankAccountDeatil().getBankAccountHolderName());
-		
-		return farmerDto;
-	}
-	
-	public List<Crop> getFarmerCrops(String userName)
-	{
-		FarmerModel farmerModel =repo.findById(userName).get();
-		return farmerModel.getCrops();
-		
+	@Override
+	public FarmerDto getFarmerDetails(String userName, ServerHttpRequest request) {
+		if (validateToken(request)) {
+			FarmerModel farmerModel = repo.findById(userName).get();
+			FarmerDto farmerDto = new FarmerDto(farmerModel.getFirstName(), farmerModel.getLastName(),
+					farmerModel.getEmail(), farmerModel.getContact(), farmerModel.getUserName(),
+					farmerModel.getPassword(), farmerModel.getRating(), farmerModel.getAddress().getHouseNo(),
+					farmerModel.getAddress().getLocality(), farmerModel.getAddress().getTown(),
+					farmerModel.getAddress().getDistrict(), farmerModel.getAddress().getState(),
+					farmerModel.getAddress().getPostalCode(), 0, null, null, 0, 0,
+					farmerModel.getBankAccountDeatil().getBankAccountNo(),
+					farmerModel.getBankAccountDeatil().getIfscCode(),
+					farmerModel.getBankAccountDeatil().getBankAccountHolderName());
+
+			return farmerDto;
+		} else
+			log.info("invalid token provided");
+			throw new TokenValidationError("Token is not valid", org.springframework.http.HttpStatus.FORBIDDEN);
 	}
 
+	@Override
+	public List<Crop> getFarmerCrops(String userName, ServerHttpRequest request) {
+		if (validateToken(request)) {
+			FarmerModel farmerModel = repo.findById(userName).get();
+			return farmerModel.getCrops();
+		} else
+			log.info("invalid token provided");
+			throw new TokenValidationError("Token is not valid", org.springframework.http.HttpStatus.FORBIDDEN);
+
+	}
+
+	@Override
 	public List<FarmerModel> getFarmers() {
-	
+
 		return repo.findAll();
+	}
+
+	@Override
+	public Boolean validateToken(ServerHttpRequest request) {
+		String authorizationHeader = request.getHeaders().getFirst("Authorization");
+
+		String username = null;
+		String jwt = null;
+
+		if (authorizationHeader != null && authorizationHeader.startsWith("Bearer ")) {
+			jwt = authorizationHeader.substring("Bearer ".length());
+			username = jwtUtil.extractUsername(jwt);
+		} else
+			return false;
+
+		if (username != null) {
+
+			FarmerModel farmerModel = repo.findById(username).get();
+			if (farmerModel != null) {
+				if (jwtUtil.validateToken(jwt, farmerModel.getUserName())) {
+
+					return true;
+
+				} else
+					return false;
+			} else
+				return false;
+		} else
+			return false;
 	}
 
 }
